example/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "fastapi>=0.128.4",
    "logfire>=4.22.0",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
    "pillow>=12.1.0",
    "pydantic>=2.12.5",
    "pydantic-ai>=1.56.0",
    "python-dotenv>=1.2.1",
    "slowapi>=0.1.9",
    "supabase>=2.27.3",
    "tavily-python>=0.7.21",
    "uvicorn>=0.40.0",
    "google-genai>=1.2.0",
]
